Add files via upload
This is a new version of the P0 Project. This time I spent some time in the DB making sure I could complete as many of  the requirements as I could from the Backend Level. From there I was able to foreshadow which requirements I could meet with the BL. 
Uploading Now in case something goes wrong and I at least have this version to fall back on.