P0 PROJECT SUBMISSION
This is about the 10th version of the project. I revamped the DB several times and created a brand new project for it every time. My goal was to meet the minimum requirements at least by today. However, I plan to add more requirements and apply better practices to this Solution. For example, exception handling, cleaner code, separate Model classes with methods that I could call from the main class, and also working with the Log file and Repository.